changed data types

+++ src/envs/brax/meta_ant.py
@@ -22,11 +22,11 @@ DEFAULT_CONTEXT = {
 }
 
 CONTEXT_BOUNDS = {
-    "joint_stiffness": (1, np.inf, int),
+    "joint_stiffness": (1, np.inf, float),
     "gravity": (-np.inf, -0.1, float),
     "friction": (-np.inf, np.inf, float),
     "angular_damping": (-np.inf, np.inf, float),
-    "actuator_strength": (1, np.inf, int),
+    "actuator_strength": (1, np.inf, float),
     "joint_angular_damping": (0, np.inf, float),
     "torso_mass": (0.1, np.inf, float),
 }
